Prevent overwrinting sync storage

+++ app/src/background.ts
@@ -37,7 +37,7 @@ messengerServer.instructions = {
 
 async function migrateConfigStorage() {
     const localConfig = await browser.storage.local.get();
-    if (Object.keys(localConfig).length > 0) {
+    if (Object.keys(localConfig).length > 0 && Object.keys(await browser.storage.sync.get()).length === 0) {
         await browser.storage.sync.set(localConfig);
         await browser.storage.local.clear();
         browser.runtime.reload();
